Seperate out functions to seperate scripts.
